--- a/library/Booster/Pattern/Unify.hs
+++ b/library/Booster/Pattern/Unify.hs
@@ -580,7 +580,7 @@ bindVariable var term = do
                     | otherwise ->
                         -- the term in the binding could be _equivalent_
                         -- (not necessarily syntactically equal) to term'
-                        enqueueRegularProblem oldTerm term
+                        addIndeterminate oldTerm term
                 Nothing -> do
                     let
                         -- apply existing substitutions to term