--- a/unit-tests/Test/Booster/Pattern/Unify.hs
+++ b/unit-tests/Test/Booster/Pattern/Unify.hs
@@ -125,9 +125,8 @@ constructors =
               z = var "Z" someSort
               t1 = app con3 [var "X" someSort, var "X" someSort]
               t2 = app con3 [y, z]
-           in test "Matching the same variable in a constructor (failing)" t1 t2 $
-                failed $
-                    VariableConflict (Variable someSort "X") y z
+           in test "Matching the same variable in a constructor (indeterminate)" t1 t2 $
+                remainder [(y, z)]
         ]
 
 functions :: TestTree