@ehildenb 2023-12-06
This may cause a lot of cases that were failing to become indeterminate now? Maybe we should post timing information here just in case.