@geo2a 2023-12-07
Yes , running the performance measurement now with `PYTEST_PARALLEL=3` with the hope for it to finish before the standup.