@geo2a 2023-12-07
We seem to be getting a 2x slowdown on some of the Kontrol proofs (with `PYTEST_PARALLEL=3`):
```
| Test                                                     | georgy-fix-rule-unification time | master-e6752d1 time | (georgy-fix-rule-unification/master-e6752d1) time
|----------------------------------------------------------|----------------------------------|---------------------|---------------------------------------------------
| ArithmeticTest.test_max1(uint256,uint256)                | 71.29                            | 102.16              | 0.697826938136257
| StoreTest.testGasLoadColdVM()                            | 39.04                            | 54.32               | 0.7187039764359352
| SetUpDeployTest.test_extcodesize()                       | 97.02                            | 132.99              | 0.7295285359801488
| StoreTest.testGasLoadWarmUp()                            | 62.57                            | 84.06               | 0.744349274327861
| AllowChangesTest.testFailAllowChangesToStorage()         | 129.43                           | 149.32              | 0.8667961425127244
| AddrTest.test_addr_true()                                | 34.86                            | 36.98               | 0.942671714440238
| AssumeTest.test_assume_true(uint256,uint256)             | 37.81                            | 39.7                | 0.9523929471032746
| ArithmeticTest.test_max2(uint256,uint256)                | 75.58                            | 79.16               | 0.9547751389590703
| BlockParamsTest.testRoll(uint256)                        | 36.45                            | 38.14               | 0.9556895647614054
| AssertTest.testFail_assert_true()                        | 39.83                            | 41.63               | 0.9567619505164544
| AccountParamsTest.test_Nonce_NonExistentAddress()        | 52.86                            | 54.84               | 0.9638949671772428
| ExpectRevertTest.testFail_expectRevert_empty()           | 43.38                            | 41.66               | 1.0412866058569372
| AssertTest.test_assert_true_branch(uint256)              | 108.43                           | 79.63               | 1.3616727364058774
| AssertTest.test_assert_true()                            | 76.61                            | 54.41               | 1.4080132328616064
| ExpectRevertTest.testFail_expectRevert_multipleReverts() | 185.24                           | 125.23              | 1.4791982751736805
| ExpectRevertTest.test_expectRevert_true()                | 157.43                           | 98.7                | 1.5950354609929078
| PlainPrankTest.test_prank_zeroAddress_true()             | 305.59                           | 150.96              | 2.0243110757816636
| ExpectRevertTest.testFail_ExpectRevert_failAndSuccess()  | 204.75                           | 100.84              | 2.0304442681475603
| ExpectRevertTest.testFail_expectRevert_false()           | 205.9                            | 100.16              | 2.0557108626198084
| PlainPrankTest.test_startPrank_true()                    | 290.19                           | 133.78              | 2.169158319629242
| PlainPrankTest.test_startPrankWithOrigin_true()          | 307.25                           | 134.29              | 2.2879588949288854
| PlainPrankTest.test_startPrank_zeroAddress_true()        | 353.61                           | 140.75              | 2.5123268206039078
| ExpectRevertTest.test_ExpectRevert_increasedDepth()      | 423.25                           | 166.35              | 2.5443342350465885
| AssertTest.testFail_expect_revert()                      | 196.44                           | 70.31               | 2.793912672450576
| ExpectRevertTest.test_expectRevert_returnValue()         | 592.58                           | 142.26              | 4.165471671587235
| ExpectRevertTest.test_expectRevert_message()             | 553.48                           | 103.39              | 5.35332237160267
| TOTAL                                                    | 4680.870000000001                | 2456.02             | 1.9058761736467948
```