@geo2a 2023-12-07
No significant difference in KEVM (ran with `PYTEST_PARALLEL=1`):
```
| Test                                     | georgy-fix-rule-unification time | master-e6752d1 time | (georgy-fix-rule-unification/master-e6752d1) time
|------------------------------------------|----------------------------------|---------------------|---------------------------------------------------
| mcd/dstoken-approve-fail-rough-spec.k    | 101.53                           | 108.03              | 0.9398315282791817
| benchmarks/ecrecover00-siginvalid-spec.k | 98.99                            | 105.12              | 0.9416856925418569
| erc20/ds/transferFrom-failure-1-a-spec.k | 99.22                            | 103.39              | 0.9596672792339684
| TOTAL                                    | 299.74                           | 316.54              | 0.9469261388766033
```